requirements.txt
python-telegram-bot>=21.0
google-genai
pdfplumber
Pillow
python-dotenv
easyocr
